api/requirements.txt
fastapi
uvicorn
requests
pydantic
groq
python-dotenv
upstash-redis
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
